--- ACPUserProfile/Assets/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a7cc2721b68ba43ff9bc761a0612b53c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
