+++ ACPUserProfile/Assets/Plugins/ACPUserProfile.cs
@@ -1,5 +1,5 @@
 /*
-Unity Plug-in v: 0.0.1
+Unity Plug-in v: 1.0.0
 Copyright 2020 Adobe. All rights reserved.
 This file is licensed to you under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License. You may obtain a copy
